Refactor Spark XxHash64 into shared header
