+++ velox/functions/sparksql/CMakeLists.txt
@@ -100,6 +100,7 @@ velox_add_library(
   UnscaledValueFunction.h
   Uuid.h
   VarcharTypeWriteSideCheck.h
+  XxHash64.h
 )
 
 velox_link_libraries(
