+++ velox/functions/sparksql/Hash.cpp
@@ -17,9 +17,10 @@
 
 #include <folly/CPortability.h>
 
-#include "velox/common/base/BitUtil.h"
 #include "velox/expression/DecodedArgs.h"
 #include "velox/functions/lib/Murmur3Hash32Base.h"
+#include "velox/functions/sparksql/XxHash64.h"
+#include "velox/type/DecimalUtil.h"
 #include "velox/vector/FlatVector.h"
 
 namespace facebook::velox::functions::sparksql {
@@ -468,153 +469,6 @@ class Murmur3HashFunction final : public exec::VectorFunction {
 
  private:
   const std::optional<int32_t> seed_;
-};
-
-class XxHash64 final {
- public:
-  using SeedType = int64_t;
-  using ReturnType = int64_t;
-
-  static uint64_t hashInt32(const int32_t input, uint64_t seed) {
-    int64_t hash = seed + PRIME64_5 + 4L;
-    hash ^= static_cast<int64_t>((input & 0xFFFFFFFFL) * PRIME64_1);
-    hash = bits::rotateLeft64(hash, 23) * PRIME64_2 + PRIME64_3;
-    return fmix(hash);
-  }
-
-  static uint64_t hashInt64(int64_t input, uint64_t seed) {
-    int64_t hash = seed + PRIME64_5 + 8L;
-    hash ^= bits::rotateLeft64(input * PRIME64_2, 31) * PRIME64_1;
-    hash = bits::rotateLeft64(hash, 27) * PRIME64_1 + PRIME64_4;
-    return fmix(hash);
-  }
-
-  // Floating point numbers are hashed as if they are integers, with
-  // -0f defined to have the same output as +0f.
-  static uint64_t hashFloat(float input, uint64_t seed) {
-    return hashInt32(
-        input == -0.f ? 0 : *reinterpret_cast<uint32_t*>(&input), seed);
-  }
-
-  static uint64_t hashDouble(double input, uint64_t seed) {
-    return hashInt64(
-        input == -0. ? 0 : *reinterpret_cast<uint64_t*>(&input), seed);
-  }
-
-  static uint64_t hashBytes(const StringView& input, uint64_t seed) {
-    const char* i = input.data();
-    const char* const end = input.data() + input.size();
-
-    uint64_t hash = hashBytesByWords(input, seed);
-    uint32_t length = input.size();
-    auto offset = i + (length & -8);
-    if (offset + 4L <= end) {
-      hash ^= (*reinterpret_cast<const uint64_t*>(offset) & 0xFFFFFFFFL) *
-          PRIME64_1;
-      hash = bits::rotateLeft64(hash, 23) * PRIME64_2 + PRIME64_3;
-      offset += 4L;
-    }
-
-    while (offset < end) {
-      hash ^= (*reinterpret_cast<const uint64_t*>(offset) & 0xFFL) * PRIME64_5;
-      hash = bits::rotateLeft64(hash, 11) * PRIME64_1;
-      offset++;
-    }
-    return fmix(hash);
-  }
-
-  static uint64_t hashLongDecimal(int128_t input, uint64_t seed) {
-    char out[sizeof(int128_t)];
-    int32_t length = DecimalUtil::toByteArray(input, out);
-    return hashBytes(StringView(out, length), seed);
-  }
-
-  static uint64_t hashTimestamp(Timestamp input, uint64_t seed) {
-    return hashInt64(input.toMicros(), seed);
-  }
-
- private:
-  static const uint64_t PRIME64_1 = 0x9E3779B185EBCA87L;
-  static const uint64_t PRIME64_2 = 0xC2B2AE3D27D4EB4FL;
-  static const uint64_t PRIME64_3 = 0x165667B19E3779F9L;
-  static const uint64_t PRIME64_4 = 0x85EBCA77C2B2AE63L;
-  static const uint64_t PRIME64_5 = 0x27D4EB2F165667C5L;
-
-  static uint64_t fmix(uint64_t hash) {
-    hash ^= hash >> 33;
-    hash *= PRIME64_2;
-    hash ^= hash >> 29;
-    hash *= PRIME64_3;
-    hash ^= hash >> 32;
-    return hash;
-  }
-
-  static uint64_t hashBytesByWords(const StringView& input, uint64_t seed) {
-    const char* i = input.data();
-    const char* const end = input.data() + input.size();
-    uint32_t length = input.size();
-    uint64_t hash;
-    if (length >= 32) {
-      uint64_t v1 = seed + PRIME64_1 + PRIME64_2;
-      uint64_t v2 = seed + PRIME64_2;
-      uint64_t v3 = seed;
-      uint64_t v4 = seed - PRIME64_1;
-      for (; i <= end - 32; i += 32) {
-        v1 = bits::rotateLeft64(
-                 v1 + (*reinterpret_cast<const uint64_t*>(i) * PRIME64_2), 31) *
-            PRIME64_1;
-        v2 = bits::rotateLeft64(
-                 v2 + (*reinterpret_cast<const uint64_t*>(i + 8) * PRIME64_2),
-                 31) *
-            PRIME64_1;
-        v3 = bits::rotateLeft64(
-                 v3 + (*reinterpret_cast<const uint64_t*>(i + 16) * PRIME64_2),
-                 31) *
-            PRIME64_1;
-        v4 = bits::rotateLeft64(
-                 v4 + (*reinterpret_cast<const uint64_t*>(i + 24) * PRIME64_2),
-                 31) *
-            PRIME64_1;
-      }
-      hash = bits::rotateLeft64(v1, 1) + bits::rotateLeft64(v2, 7) +
-          bits::rotateLeft64(v3, 12) + bits::rotateLeft64(v4, 18);
-      v1 *= PRIME64_2;
-      v1 = bits::rotateLeft64(v1, 31);
-      v1 *= PRIME64_1;
-      hash ^= v1;
-      hash = hash * PRIME64_1 + PRIME64_4;
-
-      v2 *= PRIME64_2;
-      v2 = bits::rotateLeft64(v2, 31);
-      v2 *= PRIME64_1;
-      hash ^= v2;
-      hash = hash * PRIME64_1 + PRIME64_4;
-
-      v3 *= PRIME64_2;
-      v3 = bits::rotateLeft64(v3, 31);
-      v3 *= PRIME64_1;
-      hash ^= v3;
-      hash = hash * PRIME64_1 + PRIME64_4;
-
-      v4 *= PRIME64_2;
-      v4 = bits::rotateLeft64(v4, 31);
-      v4 *= PRIME64_1;
-      hash ^= v4;
-      hash = hash * PRIME64_1 + PRIME64_4;
-    } else {
-      hash = seed + PRIME64_5;
-    }
-
-    hash += length;
-
-    for (; i <= end - 8; i += 8) {
-      hash ^= bits::rotateLeft64(
-                  *reinterpret_cast<const uint64_t*>(i) * PRIME64_2, 31) *
-          PRIME64_1;
-      hash = bits::rotateLeft64(hash, 27) * PRIME64_1 + PRIME64_4;
-    }
-    return hash;
-  }
 };
 
 class XxHash64Function final : public exec::VectorFunction {
